move stream.py to the right place
